주석 처리 추가

+++ src/main/java/com/sparta/WeatherWear/clothes/service/ClothesService.java
@@ -6,6 +6,7 @@ import com.sparta.WeatherWear.clothes.entity.Clothes;
 import com.sparta.WeatherWear.clothes.repository.ClothesRepository;
 import com.sparta.WeatherWear.global.service.ImageService;
 import com.sparta.WeatherWear.global.security.UserDetailsImpl;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -20,32 +21,30 @@ import java.util.List;
 import java.util.stream.Collectors;
 /*
 작성자 : 이승현
-사용자 관련 서비스 처리
+사용자가 가지고 있는 옷 정보를 처리하는 Service
 */
 @Service
+@RequiredArgsConstructor
 public class ClothesService {
 
     private final ClothesRepository clothesRepository;
     private final ImageService imageService;
 
-    public ClothesService(ClothesRepository clothesRepository, ImageService imageService) {
-        this.clothesRepository = clothesRepository;
-        this.imageService = imageService;
-    }
-
-    /* 옷 목록 불러오기 */
+    /* 옷 목록 불러오기 : 페이지네이션과 타입, 색상에 따라 필터링 됩니다. */
     public ResponseEntity<Page<ClothesResponseDTO>> getClotheList(UserDetailsImpl userDetails, int page, String type, String color) {
+        // 페이지네이션 8개 아이템
         Pageable pageable = PageRequest.of(page, 8);
 
+        // 사용자의 전체 옷 목록을 받아옵니다.
         List<Clothes> clothesList = clothesRepository.findByUserId(userDetails.getUser().getId());
 
-        // 필터링 적용
+        // 필터링을 적용합니다 : 없으면 전체 선택
         List<Clothes> filteredClothes = clothesList.stream()
                 .filter(clothes -> (type == null || type.trim().isEmpty() || clothes.getType().name().equalsIgnoreCase(type)))
                 .filter(clothes -> (color == null || color.trim().isEmpty() || clothes.getColor().name().equalsIgnoreCase(color)))
                 .collect(Collectors.toList());
 
-        // 페이지네이션 적용
+        // 페이지네이션을 적용합니다.
         int start = (int) pageable.getOffset();
         int end = Math.min((start + pageable.getPageSize()), filteredClothes.size());
         Page<Clothes> clothesPage = new PageImpl<>(filteredClothes.subList(start, end), pageable, filteredClothes.size());
@@ -53,8 +52,8 @@ public class ClothesService {
         return ResponseEntity.ok(clothesPage.map(ClothesResponseDTO::new));
     }
 
-    /* 옷 정보 불러오기 */
-    public ResponseEntity<ClothesResponseDTO> getClothe(UserDetailsImpl userDetails, long id) {
+    /* 단일 옷 아이템 정보를 불러오기 */
+    public ResponseEntity<ClothesResponseDTO> getClothe(long id) {
         Clothes clothes = clothesRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("No Clothes"));
         return ResponseEntity.ok(new ClothesResponseDTO(clothes));
     }
@@ -63,7 +62,9 @@ public class ClothesService {
     /* 옷 추가 */
     @Transactional
     public ResponseEntity<String> createClothes(UserDetailsImpl userDetails, ClothesRequestDTO clothesRequestDTO,MultipartFile file) throws IOException {
+        // 타입과 색상을 기준으로 옷 정보를 추가합니다.
         Clothes savedClothes = clothesRepository.save( new Clothes(clothesRequestDTO,userDetails.getUser()));
+        // 파일이 있을 경우 저장하고 옷 정보에 추가합니다.
         if(file != null){
             String imageUrl = imageService.uploadImagefile("clothes/", String.valueOf(savedClothes.getId()),file);
             savedClothes.updateImage(imageUrl);
@@ -75,12 +76,13 @@ public class ClothesService {
     /* 옷 삭제 */
     @Transactional
     public ResponseEntity<String> removeClothes(UserDetailsImpl userDetails, long id) throws IOException {
+        // 옷 객체 검색
         Clothes clothes = clothesRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("No Clothes"));
-        if(!clothes.getUser().getId().equals(userDetails.getUser().getId()))
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("사용자의 옷장 아이템이 아닙니다.");
-        if(!clothes.getImage().isEmpty()){
-            imageService.deleteImage(clothes.getImage()); // 이미지가 있으면 삭제
-        }
+        // 사용자와 옷의 등록자가 일치하는지 확인
+        if(!clothes.getUser().getId().equals(userDetails.getUser().getId()))  return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("사용자의 옷장 아이템이 아닙니다.");
+        // 옷 이미지가 있을 경우 해당 이미지 삭제
+        if(!clothes.getImage().isEmpty()) imageService.deleteImage(clothes.getImage());
+        // 삭제
         clothesRepository.delete(clothes);
         return ResponseEntity.ok("Clothes delete successfully");
     }
